requirements.txt
pandas
numpy>=1.21
matplotlib
seaborn
joblib
PyYAML
dvc
dvc[gdrive]
fastapi
scikit-learn==1.2.2
imbalanced-learn==0.10.1
pytest
httpx>=0.27.0



